requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
bs4==0.0.1
requests==2.27.1
lxml==4.8.0
azure-data-tables==12.4.0